Add OptionOnWarning for the web api client
The Slack web api returns a "warning" property on response objects. The
requests succeed, but this field will return a comma-separated list of
warning codes that range from "superfluous_charset" to
"method_deprecated", as well as extra messages in the
.response_metadata.warnings property.

This option allows applications to get notified when these warnings
exist.